requirements.txt

flet>=0.19.0

opencv-python>=4.8.1.78

rich>=13.7.0

python-i18n>=0.3.9

pyinstaller>=6.3.0

Pillow>=10.3.0 


